plan: address comment
